refactor: Разделены модели по группам Добавлены две группы - entities и forms Групп по типу queries, traits и т.д. пока нет
