+++ config/test.php
@@ -25,7 +25,7 @@ return [
             'showScriptName' => true,
         ],
         'user' => [
-            'identityClass' => 'app\models\User',
+            'identityClass' => 'app\models\entities\User',
         ],
         'request' => [
             'cookieValidationKey' => 'test',
