+++ config/web.php
@@ -20,7 +20,7 @@ $config = [
             'class' => 'yii\caching\FileCache',
         ],
         'user' => [
-            'identityClass' => 'app\models\User',
+            'identityClass' => 'app\models\entities\User',
             'enableAutoLogin' => true,
         ],
         'errorHandler' => [
