+++ models/entities/User.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace app\models;
+namespace app\models\entities;
 
 use Yii;
 use yii\base\NotSupportedException;
