+++ models/forms/PasswordResetRequestForm.php
@@ -1,9 +1,10 @@
 <?php
-namespace frontend\models;
+
+namespace app\models\forms;
 
 use Yii;
 use yii\base\Model;
-use app\models\User;
+use app\models\entities\User;
 
 /**
  * Password reset request form
@@ -23,7 +24,7 @@ class PasswordResetRequestForm extends Model
             ['email', 'required'],
             ['email', 'email'],
             ['email', 'exist',
-                'targetClass' => '\app\models\User',
+                'targetClass' => '\app\models\entities\User',
                 'filter' => ['status' => User::STATUS_ACTIVE],
                 'message' => 'There is no user with this email address.'
             ],
