+++ tests/unit/models/LoginFormTest.php
@@ -2,7 +2,7 @@
 
 namespace tests\unit\models;
 
-use app\models\LoginForm;
+use app\models\forms\LoginForm;
 
 class LoginFormTest extends \Codeception\Test\Unit
 {
