+++ views/site/contact.php
@@ -2,7 +2,7 @@
 
 /* @var $this yii\web\View */
 /* @var $form yii\bootstrap\ActiveForm */
-/* @var $model app\models\ContactForm */
+/* @var $model app\models\forms\ContactForm */
 
 use yii\helpers\Html;
 use yii\bootstrap\ActiveForm;
